Use a channel for message passing
Use a basic mpsc channel for passing information about open ports back
to the main thread. Currently only open ports, and only as integers -
in future add more detail and formatting, perhaps pass all
non-debug/trace outputting to the main thread so that we can present
info in a variety of formats.